ci: Clean up cross-browser-testing
